requirements.txt
requests
google-auth>=2.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.70.0
